fix(api): Fix bug in filesystem creation
The bug  occurred when having  multiple addons  of which only  one has
a filesystem.  When the  0th filesystem handler  supported  the wanted
filesystem, the 0th ADDON was used. That  works fine when the handlers
are identical in size to the addons  array, but that isn't necessarily
the case if there are multiple addons.